@@ -2,10 +2,10 @@ name: camera_web
 description: A Flutter plugin for getting information about and controlling the camera on Web.
 repository: https://github.com/flutter/packages/tree/main/packages/camera/camera_web
 issue_tracker: https://github.com/flutter/flutter/issues?q=is%3Aissue+is%3Aopen+label%3A%22p%3A+camera%22
-version: 0.3.4
+version: 0.3.5
 
 environment:
-  sdk: ^3.2.0
+  sdk: ^3.3.0
   flutter: ">=3.16.0"
 
 flutter:
@@ -23,6 +23,7 @@ dependencies:
   flutter_web_plugins:
     sdk: flutter
   stream_transform: ^2.0.0
+  web: ^0.5.1
 
 dev_dependencies:
   flutter_test:
