@@ -2,19 +2,23 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:html';
-import 'dart:js_util' as js_util;
+// ignore_for_file: only_throw_errors
+
+import 'dart:js_interop';
+import 'dart:js_interop_unsafe';
 
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 // ignore_for_file: implementation_imports
 import 'package:camera_web/src/camera.dart';
 import 'package:camera_web/src/camera_service.dart';
+import 'package:camera_web/src/pkg_web_tweaks.dart';
 import 'package:camera_web/src/shims/dart_js_util.dart';
 import 'package:camera_web/src/types/types.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:web/web.dart' as web;
 
 import 'helpers/helpers.dart';
 
@@ -24,27 +28,38 @@ void main() {
   group('CameraService', () {
     const int cameraId = 1;
 
-    late Window window;
-    late Navigator navigator;
-    late MediaDevices mediaDevices;
-    late CameraService cameraService;
+    late MockWindow mockWindow;
+    late MockNavigator mockNavigator;
+    late MockMediaDevices mockMediaDevices;
+
+    late web.Window window;
+    late web.Navigator navigator;
+    late web.MediaDevices mediaDevices;
+
     late JsUtil jsUtil;
 
-    setUp(() async {
-      window = MockWindow();
-      navigator = MockNavigator();
-      mediaDevices = MockMediaDevices();
-      jsUtil = MockJsUtil();
+    late CameraService cameraService;
 
-      when(() => window.navigator).thenReturn(navigator);
-      when(() => navigator.mediaDevices).thenReturn(mediaDevices);
+    setUp(() async {
+      mockWindow = MockWindow();
+      mockNavigator = MockNavigator();
+      mockMediaDevices = MockMediaDevices();
+
+      window = createJSInteropWrapper(mockWindow) as web.Window;
+      navigator = createJSInteropWrapper(mockNavigator) as web.Navigator;
+      mediaDevices =
+          createJSInteropWrapper(mockMediaDevices) as web.MediaDevices;
+
+      mockWindow.navigator = navigator;
+      mockNavigator.mediaDevices = mediaDevices;
+
+      jsUtil = MockJsUtil();
 
       // Mock JsUtil to return the real getProperty from dart:js_util.
       when<dynamic>(() => jsUtil.getProperty(any(), any())).thenAnswer(
-        (Invocation invocation) => js_util.getProperty<dynamic>(
-          invocation.positionalArguments[0] as Object,
-          invocation.positionalArguments[1] as Object,
-        ),
+        (Invocation invocation) =>
+            (invocation.positionalArguments[0] as JSObject)
+                .getProperty(invocation.positionalArguments[1] as JSAny),
       );
 
       cameraService = CameraService()..window = window;
@@ -54,8 +69,15 @@ void main() {
       testWidgets(
           'calls MediaDevices.getUserMedia '
           'with provided options', (WidgetTester tester) async {
-        when(() => mediaDevices.getUserMedia(any()))
-            .thenAnswer((_) async => FakeMediaStream(<MediaStreamTrack>[]));
+        late final web.MediaStreamConstraints? capturedConstraints;
+        mockMediaDevices.getUserMedia =
+            ([web.MediaStreamConstraints? constraints]) {
+          capturedConstraints = constraints;
+          final web.MediaStream stream =
+              createJSInteropWrapper(FakeMediaStream(<web.MediaStreamTrack>[]))
+                  as web.MediaStream;
+          return Future<web.MediaStream>.value(stream).toJS;
+        };
 
         final CameraOptions options = CameraOptions(
           video: VideoConstraints(
@@ -66,27 +88,8 @@ void main() {
 
         await cameraService.getMediaStreamForOptions(options);
 
-        verify(
-          () => mediaDevices.getUserMedia(options.toJson()),
-        ).called(1);
-      });
-
-      testWidgets(
-          'throws PlatformException '
-          'with notSupported error '
-          'when there are no media devices', (WidgetTester tester) async {
-        when(() => navigator.mediaDevices).thenReturn(null);
-
-        expect(
-          () => cameraService.getMediaStreamForOptions(const CameraOptions()),
-          throwsA(
-            isA<PlatformException>().having(
-              (PlatformException e) => e.code,
-              'code',
-              CameraErrorCode.notSupported.toString(),
-            ),
-          ),
-        );
+        expect(capturedConstraints?.video.dartify(), options.video.toJson());
+        expect(capturedConstraints?.audio.dartify(), options.audio.toJson());
       });
 
       group('throws CameraWebException', () {
@@ -94,8 +97,8 @@ void main() {
             'with notFound error '
             'when MediaDevices.getUserMedia throws DomException '
             'with NotFoundError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('NotFoundError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'NotFoundError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -116,8 +119,8 @@ void main() {
             'with notFound error '
             'when MediaDevices.getUserMedia throws DomException '
             'with DevicesNotFoundError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('DevicesNotFoundError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'DevicesNotFoundError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -138,9 +141,8 @@ void main() {
             'with notReadable error '
             'when MediaDevices.getUserMedia throws DomException '
             'with NotReadableError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('NotReadableError'));
-
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'NotReadableError');
           expect(
             () => cameraService.getMediaStreamForOptions(
               const CameraOptions(),
@@ -160,8 +162,8 @@ void main() {
             'with notReadable error '
             'when MediaDevices.getUserMedia throws DomException '
             'with TrackStartError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('TrackStartError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'TrackStartError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -182,8 +184,8 @@ void main() {
             'with overconstrained error '
             'when MediaDevices.getUserMedia throws DomException '
             'with OverconstrainedError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('OverconstrainedError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'OverconstrainedError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -204,8 +206,8 @@ void main() {
             'with overconstrained error '
             'when MediaDevices.getUserMedia throws DomException '
             'with ConstraintNotSatisfiedError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('ConstraintNotSatisfiedError'));
+          mockMediaDevices.getUserMedia = ([_]) =>
+              throw web.DOMException('', 'ConstraintNotSatisfiedError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -226,8 +228,8 @@ void main() {
             'with permissionDenied error '
             'when MediaDevices.getUserMedia throws DomException '
             'with NotAllowedError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('NotAllowedError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'NotAllowedError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -248,8 +250,8 @@ void main() {
             'with permissionDenied error '
             'when MediaDevices.getUserMedia throws DomException '
             'with PermissionDeniedError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('PermissionDeniedError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'PermissionDeniedError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -270,8 +272,8 @@ void main() {
             'with type error '
             'when MediaDevices.getUserMedia throws DomException '
             'with TypeError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('TypeError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'TypeError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -292,8 +294,8 @@ void main() {
             'with abort error '
             'when MediaDevices.getUserMedia throws DomException '
             'with AbortError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('AbortError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'AbortError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -314,8 +316,8 @@ void main() {
             'with security error '
             'when MediaDevices.getUserMedia throws DomException '
             'with SecurityError', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('SecurityError'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'SecurityError');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -336,8 +338,8 @@ void main() {
             'with unknown error '
             'when MediaDevices.getUserMedia throws DomException '
             'with an unknown error', (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any()))
-              .thenThrow(FakeDomException('Unknown'));
+          mockMediaDevices.getUserMedia =
+              ([_]) => throw web.DOMException('', 'Unknown');
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -358,7 +360,7 @@ void main() {
             'with unknown error '
             'when MediaDevices.getUserMedia throws an unknown exception',
             (WidgetTester tester) async {
-          when(() => mediaDevices.getUserMedia(any())).thenThrow(Exception());
+          mockMediaDevices.getUserMedia = ([_]) => throw Exception();
 
           expect(
             () => cameraService.getMediaStreamForOptions(
@@ -379,17 +381,23 @@ void main() {
 
     group('getZoomLevelCapabilityForCamera', () {
       late Camera camera;
-      late List<MediaStreamTrack> videoTracks;
+      late MockMediaStreamTrack mockVideoTrack;
+      late List<web.MediaStreamTrack> videoTracks;
 
       setUp(() {
         camera = MockCamera();
-        videoTracks = <MediaStreamTrack>[
-          MockMediaStreamTrack(),
-          MockMediaStreamTrack()
+        mockVideoTrack = MockMediaStreamTrack();
+        videoTracks = <web.MediaStreamTrack>[
+          createJSInteropWrapper(mockVideoTrack) as web.MediaStreamTrack,
+          createJSInteropWrapper(MockMediaStreamTrack())
+              as web.MediaStreamTrack,
         ];
 
         when(() => camera.textureId).thenReturn(0);
-        when(() => camera.stream).thenReturn(FakeMediaStream(videoTracks));
+        when(() => camera.stream).thenReturn(
+          createJSInteropWrapper(FakeMediaStream(videoTracks))
+              as web.MediaStream,
+        );
 
         cameraService.jsUtil = jsUtil;
       });
@@ -397,18 +405,15 @@ void main() {
       testWidgets(
           'returns the zoom level capability '
           'based on the first video track', (WidgetTester tester) async {
-        when(mediaDevices.getSupportedConstraints)
-            .thenReturn(<dynamic, dynamic>{
-          'zoom': true,
-        });
+        mockMediaDevices.getSupportedConstraints =
+            () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                  zoom: true,
+                );
 
-        when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-          'zoom': js_util.jsify(<dynamic, dynamic>{
-            'min': 100,
-            'max': 400,
-            'step': 2,
-          }),
-        });
+        mockVideoTrack.getCapabilities =
+            () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                  zoom: MediaSettingsRange(min: 100, max: 400, step: 2),
+                );
 
         final ZoomLevelCapability zoomLevelCapability =
             cameraService.getZoomLevelCapabilityForCamera(camera);
@@ -421,73 +426,17 @@ void main() {
       group('throws CameraWebException', () {
         testWidgets(
             'with zoomLevelNotSupported error '
-            'when there are no media devices', (WidgetTester tester) async {
-          when(() => navigator.mediaDevices).thenReturn(null);
-
-          expect(
-            () => cameraService.getZoomLevelCapabilityForCamera(camera),
-            throwsA(
-              isA<CameraWebException>()
-                  .having(
-                    (CameraWebException e) => e.cameraId,
-                    'cameraId',
-                    camera.textureId,
-                  )
-                  .having(
-                    (CameraWebException e) => e.code,
-                    'code',
-                    CameraErrorCode.zoomLevelNotSupported,
-                  ),
-            ),
-          );
-        });
-
-        testWidgets(
-            'with zoomLevelNotSupported error '
             'when the zoom level is not supported '
             'in the browser', (WidgetTester tester) async {
-          when(mediaDevices.getSupportedConstraints)
-              .thenReturn(<dynamic, dynamic>{
-            'zoom': false,
-          });
+          mockMediaDevices.getSupportedConstraints =
+              () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                    zoom: false,
+                  );
 
-          when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-            'zoom': <dynamic, dynamic>{
-              'min': 100,
-              'max': 400,
-              'step': 2,
-            },
-          });
-
-          expect(
-            () => cameraService.getZoomLevelCapabilityForCamera(camera),
-            throwsA(
-              isA<CameraWebException>()
-                  .having(
-                    (CameraWebException e) => e.cameraId,
-                    'cameraId',
-                    camera.textureId,
-                  )
-                  .having(
-                    (CameraWebException e) => e.code,
-                    'code',
-                    CameraErrorCode.zoomLevelNotSupported,
-                  ),
-            ),
-          );
-        });
-
-        testWidgets(
-            'with zoomLevelNotSupported error '
-            'when the zoom level is not supported '
-            'by the camera', (WidgetTester tester) async {
-          when(mediaDevices.getSupportedConstraints)
-              .thenReturn(<dynamic, dynamic>{
-            'zoom': true,
-          });
-
-          when(videoTracks.first.getCapabilities)
-              .thenReturn(<dynamic, dynamic>{});
+          mockVideoTrack.getCapabilities =
+              () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                    zoom: MediaSettingsRange(min: 100, max: 400, step: 2),
+                  );
 
           expect(
             () => cameraService.getZoomLevelCapabilityForCamera(camera),
@@ -511,14 +460,16 @@ void main() {
             'with notStarted error '
             'when the camera stream has not been initialized',
             (WidgetTester tester) async {
-          when(mediaDevices.getSupportedConstraints)
-              .thenReturn(<dynamic, dynamic>{
-            'zoom': true,
-          });
+          mockMediaDevices.getSupportedConstraints =
+              () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                    zoom: true,
+                  );
 
           // Create a camera stream with no video tracks.
-          when(() => camera.stream)
-              .thenReturn(FakeMediaStream(<MediaStreamTrack>[]));
+          when(() => camera.stream).thenReturn(
+            createJSInteropWrapper(FakeMediaStream(<web.MediaStreamTrack>[]))
+                as web.MediaStream,
+          );
 
           expect(
             () => cameraService.getZoomLevelCapabilityForCamera(camera),
@@ -546,58 +497,40 @@ void main() {
       });
 
       testWidgets(
-          'throws PlatformException '
-          'with notSupported error '
-          'when there are no media devices', (WidgetTester tester) async {
-        when(() => navigator.mediaDevices).thenReturn(null);
-
-        expect(
-          () =>
-              cameraService.getFacingModeForVideoTrack(MockMediaStreamTrack()),
-          throwsA(
-            isA<PlatformException>().having(
-              (PlatformException e) => e.code,
-              'code',
-              CameraErrorCode.notSupported.toString(),
-            ),
-          ),
-        );
-      });
-
-      testWidgets(
           'returns null '
           'when the facing mode is not supported', (WidgetTester tester) async {
-        when(mediaDevices.getSupportedConstraints)
-            .thenReturn(<dynamic, dynamic>{
-          'facingMode': false,
-        });
+        mockMediaDevices.getSupportedConstraints =
+            () => web.MediaTrackSupportedConstraints(facingMode: false);
 
-        final String? facingMode =
-            cameraService.getFacingModeForVideoTrack(MockMediaStreamTrack());
+        final String? facingMode = cameraService.getFacingModeForVideoTrack(
+          createJSInteropWrapper(MockMediaStreamTrack())
+              as web.MediaStreamTrack,
+        );
 
         expect(facingMode, isNull);
       });
 
       group('when the facing mode is supported', () {
-        late MediaStreamTrack videoTrack;
+        late MockMediaStreamTrack mockVideoTrack;
+        late web.MediaStreamTrack videoTrack;
 
         setUp(() {
-          videoTrack = MockMediaStreamTrack();
+          mockVideoTrack = MockMediaStreamTrack();
+          videoTrack =
+              createJSInteropWrapper(mockVideoTrack) as web.MediaStreamTrack;
 
-          when(() => jsUtil.hasProperty(videoTrack, 'getCapabilities'))
+          when(() => jsUtil.hasProperty(videoTrack, 'getCapabilities'.toJS))
               .thenReturn(true);
 
-          when(mediaDevices.getSupportedConstraints)
-              .thenReturn(<dynamic, dynamic>{
-            'facingMode': true,
-          });
+          mockMediaDevices.getSupportedConstraints =
+              () => web.MediaTrackSupportedConstraints(facingMode: true);
         });
 
         testWidgets(
             'returns an appropriate facing mode '
             'based on the video track settings', (WidgetTester tester) async {
-          when(videoTrack.getSettings)
-              .thenReturn(<dynamic, dynamic>{'facingMode': 'user'});
+          mockVideoTrack.getSettings =
+              () => web.MediaTrackSettings(facingMode: 'user');
 
           final String? facingMode =
               cameraService.getFacingModeForVideoTrack(videoTrack);
@@ -610,12 +543,13 @@ void main() {
             'based on the video track capabilities '
             'when the facing mode setting is empty',
             (WidgetTester tester) async {
-          when(videoTrack.getSettings).thenReturn(<dynamic, dynamic>{});
-          when(videoTrack.getCapabilities).thenReturn(<dynamic, dynamic>{
-            'facingMode': <dynamic>['environment', 'left']
-          });
+          mockVideoTrack.getSettings =
+              () => web.MediaTrackSettings(facingMode: '');
+          mockVideoTrack.getCapabilities = () => web.MediaTrackCapabilities(
+                facingMode: <JSString>['environment'.toJS, 'left'.toJS].toJS,
+              );
 
-          when(() => jsUtil.hasProperty(videoTrack, 'getCapabilities'))
+          when(() => jsUtil.hasProperty(videoTrack, 'getCapabilities'.toJS))
               .thenReturn(true);
 
           final String? facingMode =
@@ -628,9 +562,10 @@ void main() {
             'returns null '
             'when the facing mode setting '
             'and capabilities are empty', (WidgetTester tester) async {
-          when(videoTrack.getSettings).thenReturn(<dynamic, dynamic>{});
-          when(videoTrack.getCapabilities)
-              .thenReturn(<dynamic, dynamic>{'facingMode': <dynamic>[]});
+          mockVideoTrack.getSettings =
+              () => web.MediaTrackSettings(facingMode: '');
+          mockVideoTrack.getCapabilities =
+              () => web.MediaTrackCapabilities(facingMode: <JSString>[].toJS);
 
           final String? facingMode =
               cameraService.getFacingModeForVideoTrack(videoTrack);
@@ -643,9 +578,10 @@ void main() {
             'when the facing mode setting is empty and '
             'the video track capabilities are not supported',
             (WidgetTester tester) async {
-          when(videoTrack.getSettings).thenReturn(<dynamic, dynamic>{});
+          mockVideoTrack.getSettings =
+              () => web.MediaTrackSettings(facingMode: '');
 
-          when(() => jsUtil.hasProperty(videoTrack, 'getCapabilities'))
+          when(() => jsUtil.hasProperty(videoTrack, 'getCapabilities'.toJS))
               .thenReturn(false);
 
           final String? facingMode =
