@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:html';
+import 'dart:js_interop';
 import 'dart:ui';
 
 import 'package:async/async.dart';
@@ -11,10 +11,12 @@ import 'package:camera_platform_interface/camera_platform_interface.dart';
 // ignore_for_file: implementation_imports
 import 'package:camera_web/src/camera.dart';
 import 'package:camera_web/src/camera_service.dart';
+import 'package:camera_web/src/pkg_web_tweaks.dart';
 import 'package:camera_web/src/types/types.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:mocktail/mocktail.dart';
+import 'package:web/web.dart';
 
 import 'helpers/helpers.dart';
 
@@ -24,6 +26,10 @@ void main() {
   group('Camera', () {
     const int textureId = 1;
 
+    late MockWindow mockWindow;
+    late MockNavigator mockNavigator;
+    late MockMediaDevices mockMediaDevices;
+
     late Window window;
     late Navigator navigator;
     late MediaDevices mediaDevices;
@@ -32,16 +38,20 @@ void main() {
     late CameraService cameraService;
 
     setUp(() {
-      window = MockWindow();
-      navigator = MockNavigator();
-      mediaDevices = MockMediaDevices();
+      mockWindow = MockWindow();
+      mockNavigator = MockNavigator();
+      mockMediaDevices = MockMediaDevices();
 
-      when(() => window.navigator).thenReturn(navigator);
-      when(() => navigator.mediaDevices).thenReturn(mediaDevices);
+      window = createJSInteropWrapper(mockWindow) as Window;
+      navigator = createJSInteropWrapper(mockNavigator) as Navigator;
+      mediaDevices = createJSInteropWrapper(mockMediaDevices) as MediaDevices;
+
+      mockWindow.navigator = navigator;
+      mockNavigator.mediaDevices = mediaDevices;
 
       cameraService = MockCameraService();
 
-      final VideoElement videoElement =
+      final HTMLVideoElement videoElement =
           getVideoElementWithBlankStream(const Size(10, 10));
       mediaStream = videoElement.captureStream();
 
@@ -110,7 +120,8 @@ void main() {
         expect(camera.videoElement.autoplay, isFalse);
         expect(camera.videoElement.muted, isTrue);
         expect(camera.videoElement.srcObject, mediaStream);
-        expect(camera.videoElement.attributes.keys, contains('playsinline'));
+        expect(camera.videoElement.attributes.getNamedItem('playsinline'),
+            isNotNull);
 
         expect(
             camera.videoElement.style.transformOrigin, equals('center center'));
@@ -303,26 +314,27 @@ void main() {
       group(
           'enables the torch mode '
           'when taking a picture', () {
+        late MockMediaStreamTrack mockVideoTrack;
         late List<MediaStreamTrack> videoTracks;
         late MediaStream videoStream;
-        late VideoElement videoElement;
+        late HTMLVideoElement videoElement;
 
         setUp(() {
+          mockVideoTrack = MockMediaStreamTrack();
           videoTracks = <MediaStreamTrack>[
-            MockMediaStreamTrack(),
-            MockMediaStreamTrack()
+            createJSInteropWrapper(mockVideoTrack) as MediaStreamTrack,
+            createJSInteropWrapper(MockMediaStreamTrack()) as MediaStreamTrack,
           ];
-          videoStream = FakeMediaStream(videoTracks);
+          videoStream = createJSInteropWrapper(FakeMediaStream(videoTracks))
+              as MediaStream;
 
           videoElement = getVideoElementWithBlankStream(const Size(100, 100))
             ..muted = true;
 
-          when(() => videoTracks.first.applyConstraints(any()))
-              .thenAnswer((_) async => <dynamic, dynamic>{});
-
-          when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-            'torch': true,
-          });
+          mockVideoTrack.getCapabilities =
+              () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                    torch: <JSBoolean>[true.toJS].toJS,
+                  );
         });
 
         testWidgets('if the flash mode is auto', (WidgetTester tester) async {
@@ -337,27 +349,22 @@ void main() {
 
           await camera.play();
 
+          final List<MediaTrackConstraints> capturedConstraints =
+              <MediaTrackConstraints>[];
+          mockVideoTrack.applyConstraints = ([
+            MediaTrackConstraints? constraints,
+          ]) {
+            if (constraints != null) {
+              capturedConstraints.add(constraints);
+            }
+            return Future<JSAny?>.value().toJS;
+          };
+
           final XFile _ = await camera.takePicture();
 
-          verify(
-            () => videoTracks.first.applyConstraints(<dynamic, dynamic>{
-              'advanced': <dynamic>[
-                <dynamic, dynamic>{
-                  'torch': true,
-                }
-              ]
-            }),
-          ).called(1);
-
-          verify(
-            () => videoTracks.first.applyConstraints(<dynamic, dynamic>{
-              'advanced': <dynamic>[
-                <dynamic, dynamic>{
-                  'torch': false,
-                }
-              ]
-            }),
-          ).called(1);
+          expect(capturedConstraints.length, 2);
+          expect(capturedConstraints[0].torch, true);
+          expect(capturedConstraints[1].torch, false);
         });
 
         testWidgets('if the flash mode is always', (WidgetTester tester) async {
@@ -372,27 +379,22 @@ void main() {
 
           await camera.play();
 
+          final List<MediaTrackConstraints> capturedConstraints =
+              <MediaTrackConstraints>[];
+          mockVideoTrack.applyConstraints = ([
+            MediaTrackConstraints? constraints,
+          ]) {
+            if (constraints != null) {
+              capturedConstraints.add(constraints);
+            }
+            return Future<JSAny?>.value().toJS;
+          };
+
           final XFile _ = await camera.takePicture();
 
-          verify(
-            () => videoTracks.first.applyConstraints(<dynamic, dynamic>{
-              'advanced': <dynamic>[
-                <dynamic, dynamic>{
-                  'torch': true,
-                }
-              ]
-            }),
-          ).called(1);
-
-          verify(
-            () => videoTracks.first.applyConstraints(<dynamic, dynamic>{
-              'advanced': <dynamic>[
-                <dynamic, dynamic>{
-                  'torch': false,
-                }
-              ]
-            }),
-          ).called(1);
+          expect(capturedConstraints.length, 2);
+          expect(capturedConstraints[0].torch, true);
+          expect(capturedConstraints[1].torch, false);
         });
       });
     });
@@ -404,7 +406,7 @@ void main() {
           (WidgetTester tester) async {
         const Size videoSize = Size(1280, 720);
 
-        final VideoElement videoElement =
+        final HTMLVideoElement videoElement =
             getVideoElementWithBlankStream(videoSize);
         mediaStream = videoElement.captureStream();
 
@@ -425,7 +427,7 @@ void main() {
           'returns Size.zero '
           'if the camera is missing video tracks', (WidgetTester tester) async {
         // Create a video stream with no video tracks.
-        final VideoElement videoElement = VideoElement();
+        final HTMLVideoElement videoElement = HTMLVideoElement();
         mediaStream = videoElement.captureStream();
 
         final Camera camera = Camera(
@@ -443,32 +445,37 @@ void main() {
     });
 
     group('setFlashMode', () {
+      late MockMediaStreamTrack mockVideoTrack;
       late List<MediaStreamTrack> videoTracks;
       late MediaStream videoStream;
 
       setUp(() {
+        mockVideoTrack = MockMediaStreamTrack();
         videoTracks = <MediaStreamTrack>[
-          MockMediaStreamTrack(),
-          MockMediaStreamTrack()
+          createJSInteropWrapper(mockVideoTrack) as MediaStreamTrack,
+          createJSInteropWrapper(MockMediaStreamTrack()) as MediaStreamTrack,
         ];
-        videoStream = FakeMediaStream(videoTracks);
+        videoStream =
+            createJSInteropWrapper(FakeMediaStream(videoTracks)) as MediaStream;
 
-        when(() => videoTracks.first.applyConstraints(any()))
-            .thenAnswer((_) async => <dynamic, dynamic>{});
+        mockVideoTrack.applyConstraints = ([
+          MediaTrackConstraints? constraints,
+        ]) =>
+            Future<JSAny?>.value().toJS;
 
-        when(videoTracks.first.getCapabilities)
-            .thenReturn(<dynamic, dynamic>{});
+        mockVideoTrack.getCapabilities = () => MediaTrackCapabilities();
       });
 
       testWidgets('sets the camera flash mode', (WidgetTester tester) async {
-        when(mediaDevices.getSupportedConstraints)
-            .thenReturn(<dynamic, dynamic>{
-          'torch': true,
-        });
+        mockMediaDevices.getSupportedConstraints =
+            () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                  torch: true,
+                );
 
-        when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-          'torch': true,
-        });
+        mockVideoTrack.getCapabilities =
+            () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                  torch: <JSBoolean>[true.toJS].toJS,
+                );
 
         final Camera camera = Camera(
           textureId: textureId,
@@ -490,14 +497,15 @@ void main() {
       testWidgets(
           'enables the torch mode '
           'if the flash mode is torch', (WidgetTester tester) async {
-        when(mediaDevices.getSupportedConstraints)
-            .thenReturn(<dynamic, dynamic>{
-          'torch': true,
-        });
+        mockMediaDevices.getSupportedConstraints =
+            () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                  torch: true,
+                );
 
-        when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-          'torch': true,
-        });
+        mockVideoTrack.getCapabilities =
+            () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                  torch: <JSBoolean>[true.toJS].toJS,
+                );
 
         final Camera camera = Camera(
           textureId: textureId,
@@ -506,30 +514,35 @@ void main() {
           ..window = window
           ..stream = videoStream;
 
+        final List<MediaTrackConstraints> capturedConstraints =
+            <MediaTrackConstraints>[];
+        mockVideoTrack.applyConstraints = ([
+          MediaTrackConstraints? constraints,
+        ]) {
+          if (constraints != null) {
+            capturedConstraints.add(constraints);
+          }
+          return Future<JSAny?>.value().toJS;
+        };
+
         camera.setFlashMode(FlashMode.torch);
 
-        verify(
-          () => videoTracks.first.applyConstraints(<dynamic, dynamic>{
-            'advanced': <dynamic>[
-              <dynamic, dynamic>{
-                'torch': true,
-              }
-            ]
-          }),
-        ).called(1);
+        expect(capturedConstraints.length, 1);
+        expect(capturedConstraints[0].torch, true);
       });
 
       testWidgets(
           'disables the torch mode '
           'if the flash mode is not torch', (WidgetTester tester) async {
-        when(mediaDevices.getSupportedConstraints)
-            .thenReturn(<dynamic, dynamic>{
-          'torch': true,
-        });
+        mockMediaDevices.getSupportedConstraints =
+            () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                  torch: true,
+                );
 
-        when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-          'torch': true,
-        });
+        mockVideoTrack.getCapabilities =
+            () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                  torch: <JSBoolean>[true.toJS].toJS,
+                );
 
         final Camera camera = Camera(
           textureId: textureId,
@@ -538,62 +551,37 @@ void main() {
           ..window = window
           ..stream = videoStream;
 
+        final List<MediaTrackConstraints> capturedConstraints =
+            <MediaTrackConstraints>[];
+        mockVideoTrack.applyConstraints = ([
+          MediaTrackConstraints? constraints,
+        ]) {
+          if (constraints != null) {
+            capturedConstraints.add(constraints);
+          }
+          return Future<JSAny?>.value().toJS;
+        };
+
         camera.setFlashMode(FlashMode.auto);
 
-        verify(
-          () => videoTracks.first.applyConstraints(<dynamic, dynamic>{
-            'advanced': <dynamic>[
-              <dynamic, dynamic>{
-                'torch': false,
-              }
-            ]
-          }),
-        ).called(1);
+        expect(capturedConstraints.length, 1);
+        expect(capturedConstraints[0].torch, false);
       });
 
       group('throws a CameraWebException', () {
         testWidgets(
             'with torchModeNotSupported error '
-            'when there are no media devices', (WidgetTester tester) async {
-          when(() => navigator.mediaDevices).thenReturn(null);
-
-          final Camera camera = Camera(
-            textureId: textureId,
-            cameraService: cameraService,
-          )
-            ..window = window
-            ..stream = videoStream;
-
-          expect(
-            () => camera.setFlashMode(FlashMode.always),
-            throwsA(
-              isA<CameraWebException>()
-                  .having(
-                    (CameraWebException e) => e.cameraId,
-                    'cameraId',
-                    textureId,
-                  )
-                  .having(
-                    (CameraWebException e) => e.code,
-                    'code',
-                    CameraErrorCode.torchModeNotSupported,
-                  ),
-            ),
-          );
-        });
-
-        testWidgets(
-            'with torchModeNotSupported error '
             'when the torch mode is not supported '
             'in the browser', (WidgetTester tester) async {
-          when(mediaDevices.getSupportedConstraints)
-              .thenReturn(<dynamic, dynamic>{
-            'torch': false,
-          });
+          mockMediaDevices.getSupportedConstraints =
+              () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                    torch: false,
+                  );
 
-          when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-            'torch': true,
-          });
+          mockVideoTrack.getCapabilities =
+              () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                    torch: <JSBoolean>[true.toJS].toJS,
+                  );
 
           final Camera camera = Camera(
             textureId: textureId,
@@ -624,14 +612,15 @@ void main() {
             'with torchModeNotSupported error '
             'when the torch mode is not supported '
             'by the camera', (WidgetTester tester) async {
-          when(mediaDevices.getSupportedConstraints)
-              .thenReturn(<dynamic, dynamic>{
-            'torch': true,
-          });
+          mockMediaDevices.getSupportedConstraints =
+              () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                    torch: true,
+                  );
 
-          when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-            'torch': false,
-          });
+          mockVideoTrack.getCapabilities =
+              () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                    torch: <JSBoolean>[false.toJS].toJS,
+                  );
 
           final Camera camera = Camera(
             textureId: textureId,
@@ -662,14 +651,15 @@ void main() {
             'with notStarted error '
             'when the camera stream has not been initialized',
             (WidgetTester tester) async {
-          when(mediaDevices.getSupportedConstraints)
-              .thenReturn(<dynamic, dynamic>{
-            'torch': true,
-          });
+          mockMediaDevices.getSupportedConstraints =
+              () => NonStandardFieldsOnMediaTrackSupportedConstraints.construct(
+                    torch: true,
+                  );
 
-          when(videoTracks.first.getCapabilities).thenReturn(<dynamic, dynamic>{
-            'torch': true,
-          });
+          mockVideoTrack.getCapabilities =
+              () => NonStandardFieldsOnMediaTrackCapabilities.construct(
+                    torch: <JSBoolean>[true.toJS].toJS,
+                  );
 
           final Camera camera = Camera(
             textureId: textureId,
@@ -710,7 +700,8 @@ void main() {
           final ZoomLevelCapability zoomLevelCapability = ZoomLevelCapability(
             minimum: 50.0,
             maximum: 100.0,
-            videoTrack: MockMediaStreamTrack(),
+            videoTrack: createJSInteropWrapper(MockMediaStreamTrack())
+                as MediaStreamTrack,
           );
 
           when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
@@ -741,7 +732,8 @@ void main() {
           final ZoomLevelCapability zoomLevelCapability = ZoomLevelCapability(
             minimum: 50.0,
             maximum: 100.0,
-            videoTrack: MockMediaStreamTrack(),
+            videoTrack: createJSInteropWrapper(MockMediaStreamTrack())
+                as MediaStreamTrack,
           );
 
           when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
@@ -769,7 +761,9 @@ void main() {
             cameraService: cameraService,
           );
 
-          final MockMediaStreamTrack videoTrack = MockMediaStreamTrack();
+          final MockMediaStreamTrack mockVideoTrack = MockMediaStreamTrack();
+          final MediaStreamTrack videoTrack =
+              createJSInteropWrapper(mockVideoTrack) as MediaStreamTrack;
 
           final ZoomLevelCapability zoomLevelCapability = ZoomLevelCapability(
             minimum: 50.0,
@@ -777,8 +771,16 @@ void main() {
             videoTrack: videoTrack,
           );
 
-          when(() => videoTrack.applyConstraints(any()))
-              .thenAnswer((_) async {});
+          final List<MediaTrackConstraints> capturedConstraints =
+              <MediaTrackConstraints>[];
+          mockVideoTrack.applyConstraints = ([
+            MediaTrackConstraints? constraints,
+          ]) {
+            if (constraints != null) {
+              capturedConstraints.add(constraints);
+            }
+            return Future<JSAny?>.value().toJS;
+          };
 
           when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
               .thenReturn(zoomLevelCapability);
@@ -787,15 +789,8 @@ void main() {
 
           camera.setZoomLevel(zoom);
 
-          verify(
-            () => videoTrack.applyConstraints(<dynamic, dynamic>{
-              'advanced': <dynamic>[
-                <dynamic, dynamic>{
-                  ZoomLevelCapability.constraintName: zoom,
-                }
-              ]
-            }),
-          ).called(1);
+          expect(capturedConstraints.length, 1);
+          expect(capturedConstraints[0].zoom, zoom);
         });
 
         group('throws a CameraWebException', () {
@@ -811,7 +806,8 @@ void main() {
             final ZoomLevelCapability zoomLevelCapability = ZoomLevelCapability(
               minimum: 50.0,
               maximum: 100.0,
-              videoTrack: MockMediaStreamTrack(),
+              videoTrack: createJSInteropWrapper(MockMediaStreamTrack())
+                  as MediaStreamTrack,
             );
 
             when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
@@ -846,7 +842,8 @@ void main() {
             final ZoomLevelCapability zoomLevelCapability = ZoomLevelCapability(
               minimum: 50.0,
               maximum: 100.0,
-              videoTrack: MockMediaStreamTrack(),
+              videoTrack: createJSInteropWrapper(MockMediaStreamTrack())
+                  as MediaStreamTrack,
             );
 
             when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
@@ -878,7 +875,9 @@ void main() {
           'returns a lens direction '
           'based on the first video track settings',
           (WidgetTester tester) async {
-        final MockVideoElement videoElement = MockVideoElement();
+        final MockVideoElement mockVideoElement = MockVideoElement();
+        final HTMLVideoElement videoElement =
+            createJSInteropWrapper(mockVideoElement) as HTMLVideoElement;
 
         final Camera camera = Camera(
           textureId: textureId,
@@ -887,15 +886,18 @@ void main() {
 
         final MockMediaStreamTrack firstVideoTrack = MockMediaStreamTrack();
 
-        when(() => videoElement.srcObject).thenReturn(
-          FakeMediaStream(<MediaStreamTrack>[
-            firstVideoTrack,
-            MockMediaStreamTrack(),
-          ]),
-        );
+        mockVideoElement.srcObject = createJSInteropWrapper(
+          FakeMediaStream(
+            <MediaStreamTrack>[
+              createJSInteropWrapper(firstVideoTrack) as MediaStreamTrack,
+              createJSInteropWrapper(MockMediaStreamTrack())
+                  as MediaStreamTrack,
+            ],
+          ),
+        ) as MediaStream;
 
-        when(firstVideoTrack.getSettings)
-            .thenReturn(<dynamic, dynamic>{'facingMode': 'environment'});
+        firstVideoTrack.getSettings =
+            () => MediaTrackSettings(facingMode: 'environment');
 
         when(() => cameraService.mapFacingModeToLensDirection('environment'))
             .thenReturn(CameraLensDirection.external);
@@ -910,7 +912,9 @@ void main() {
           'returns null '
           'if the first video track is missing the facing mode',
           (WidgetTester tester) async {
-        final MockVideoElement videoElement = MockVideoElement();
+        final MockVideoElement mockVideoElement = MockVideoElement();
+        final HTMLVideoElement videoElement =
+            createJSInteropWrapper(mockVideoElement) as HTMLVideoElement;
 
         final Camera camera = Camera(
           textureId: textureId,
@@ -919,14 +923,17 @@ void main() {
 
         final MockMediaStreamTrack firstVideoTrack = MockMediaStreamTrack();
 
-        when(() => videoElement.srcObject).thenReturn(
-          FakeMediaStream(<MediaStreamTrack>[
-            firstVideoTrack,
-            MockMediaStreamTrack(),
-          ]),
-        );
+        videoElement.srcObject = createJSInteropWrapper(
+          FakeMediaStream(
+            <MediaStreamTrack>[
+              createJSInteropWrapper(firstVideoTrack) as MediaStreamTrack,
+              createJSInteropWrapper(MockMediaStreamTrack())
+                  as MediaStreamTrack,
+            ],
+          ),
+        ) as MediaStream;
 
-        when(firstVideoTrack.getSettings).thenReturn(<dynamic, dynamic>{});
+        firstVideoTrack.getSettings = () => MediaTrackSettings();
 
         expect(
           camera.getLensDirection(),
@@ -938,7 +945,7 @@ void main() {
           'returns null '
           'if the camera is missing video tracks', (WidgetTester tester) async {
         // Create a video stream with no video tracks.
-        final VideoElement videoElement = VideoElement();
+        final HTMLVideoElement videoElement = HTMLVideoElement();
         mediaStream = videoElement.captureStream();
 
         final Camera camera = Camera(
@@ -974,15 +981,15 @@ void main() {
     group('video recording', () {
       const String supportedVideoType = 'video/webm';
 
+      late MockMediaRecorder mockMediaRecorder;
       late MediaRecorder mediaRecorder;
 
       bool isVideoTypeSupported(String type) => type == supportedVideoType;
 
       setUp(() {
-        mediaRecorder = MockMediaRecorder();
-
-        when(() => mediaRecorder.onError)
-            .thenAnswer((_) => const Stream<Event>.empty());
+        mockMediaRecorder = MockMediaRecorder();
+        mediaRecorder =
+            createJSInteropWrapper(mockMediaRecorder) as MediaRecorder;
       });
 
       group('startVideoRecording', () {
@@ -1027,11 +1034,21 @@ void main() {
           await camera.initialize();
           await camera.play();
 
+          final List<String> capturedEvents = <String>[];
+          mockMediaRecorder.addEventListener = (
+            String type,
+            EventListener? callback, [
+            JSAny? options,
+          ]) {
+            capturedEvents.add(type);
+          };
+
           await camera.startVideoRecording();
 
-          verify(
-            () => mediaRecorder.addEventListener('dataavailable', any()),
-          ).called(1);
+          expect(
+            capturedEvents.where((String e) => e == 'dataavailable').length,
+            1,
+          );
         });
 
         testWidgets('listens to the media recorder stop events',
@@ -1046,11 +1063,21 @@ void main() {
           await camera.initialize();
           await camera.play();
 
+          final List<String> capturedEvents = <String>[];
+          mockMediaRecorder.addEventListener = (
+            String type,
+            EventListener? callback, [
+            JSAny? options,
+          ]) {
+            capturedEvents.add(type);
+          };
+
           await camera.startVideoRecording();
 
-          verify(
-            () => mediaRecorder.addEventListener('stop', any()),
-          ).called(1);
+          expect(
+            capturedEvents.where((String e) => e == 'stop').length,
+            1,
+          );
         });
 
         testWidgets('starts a video recording', (WidgetTester tester) async {
@@ -1064,9 +1091,14 @@ void main() {
           await camera.initialize();
           await camera.play();
 
+          final List<int?> capturedStarts = <int?>[];
+          mockMediaRecorder.start = ([int? timeslice]) {
+            capturedStarts.add(timeslice);
+          };
+
           await camera.startVideoRecording();
 
-          verify(mediaRecorder.start).called(1);
+          expect(capturedStarts.length, 1);
         });
 
         group('throws a CameraWebException', () {
@@ -1108,9 +1140,14 @@ void main() {
             cameraService: cameraService,
           )..mediaRecorder = mediaRecorder;
 
+          int pauses = 0;
+          mockMediaRecorder.pause = () {
+            pauses++;
+          };
+
           await camera.pauseVideoRecording();
 
-          verify(mediaRecorder.pause).called(1);
+          expect(pauses, 1);
         });
 
         testWidgets(
@@ -1149,9 +1186,14 @@ void main() {
             cameraService: cameraService,
           )..mediaRecorder = mediaRecorder;
 
+          int resumes = 0;
+          mockMediaRecorder.resume = () {
+            resumes++;
+          };
+
           await camera.resumeVideoRecording();
 
-          verify(mediaRecorder.resume).called(1);
+          expect(resumes, 1);
         });
 
         testWidgets(
@@ -1198,50 +1240,62 @@ void main() {
           await camera.initialize();
           await camera.play();
 
-          late void Function(Event) videoDataAvailableListener;
-          late void Function(Event) videoRecordingStoppedListener;
+          late EventListener videoDataAvailableListener;
+          late EventListener videoRecordingStoppedListener;
 
-          when(
-            () => mediaRecorder.addEventListener('dataavailable', any()),
-          ).thenAnswer((Invocation invocation) {
-            videoDataAvailableListener =
-                invocation.positionalArguments[1] as void Function(Event);
-          });
-
-          when(
-            () => mediaRecorder.addEventListener('stop', any()),
-          ).thenAnswer((Invocation invocation) {
-            videoRecordingStoppedListener =
-                invocation.positionalArguments[1] as void Function(Event);
-          });
+          mockMediaRecorder.addEventListener = (
+            String type,
+            EventListener? callback, [
+            JSAny? options,
+          ]) {
+            if (type == 'dataavailable') {
+              videoDataAvailableListener = callback!;
+            } else if (type == 'stop') {
+              videoRecordingStoppedListener = callback!;
+            }
+          };
 
           Blob? finalVideo;
           List<Blob>? videoParts;
           camera.blobBuilder = (List<Blob> blobs, String videoType) {
             videoParts = <Blob>[...blobs];
-            finalVideo = Blob(blobs, videoType);
+            finalVideo = Blob(blobs.toJS, BlobPropertyBag(type: videoType));
             return finalVideo!;
           };
 
           await camera.startVideoRecording();
+
+          int stops = 0;
+          mockMediaRecorder.stop = () {
+            stops++;
+          };
+
           final Future<XFile> videoFileFuture = camera.stopVideoRecording();
 
-          final Blob capturedVideoPartOne = Blob(<Object>[]);
-          final Blob capturedVideoPartTwo = Blob(<Object>[]);
+          final Blob capturedVideoPartOne = Blob(<JSAny>[].toJS);
+          final Blob capturedVideoPartTwo = Blob(<JSAny>[].toJS);
 
           final List<Blob> capturedVideoParts = <Blob>[
             capturedVideoPartOne,
             capturedVideoPartTwo,
           ];
 
-          videoDataAvailableListener(FakeBlobEvent(capturedVideoPartOne));
-          videoDataAvailableListener(FakeBlobEvent(capturedVideoPartTwo));
+          videoDataAvailableListener.callAsFunction(
+            null,
+            createJSInteropWrapper(FakeBlobEvent(capturedVideoPartOne))
+                as BlobEvent,
+          );
+          videoDataAvailableListener.callAsFunction(
+            null,
+            createJSInteropWrapper(FakeBlobEvent(capturedVideoPartTwo))
+                as BlobEvent,
+          );
 
-          videoRecordingStoppedListener(Event('stop'));
+          videoRecordingStoppedListener.callAsFunction(null, Event('stop'));
 
           final XFile videoFile = await videoFileFuture;
 
-          verify(mediaRecorder.stop).called(1);
+          expect(stops, 1);
 
           expect(
             videoFile,
@@ -1294,15 +1348,18 @@ void main() {
       });
 
       group('on video recording stopped', () {
-        late void Function(Event) videoRecordingStoppedListener;
+        late EventListener videoRecordingStoppedListener;
 
         setUp(() {
-          when(
-            () => mediaRecorder.addEventListener('stop', any()),
-          ).thenAnswer((Invocation invocation) {
-            videoRecordingStoppedListener =
-                invocation.positionalArguments[1] as void Function(Event);
-          });
+          mockMediaRecorder.addEventListener = (
+            String type,
+            EventListener? callback, [
+            JSAny? options,
+          ]) {
+            if (type == 'stop') {
+              videoRecordingStoppedListener = callback!;
+            }
+          };
         });
 
         testWidgets('stops listening to the media recorder data events',
@@ -1319,13 +1376,23 @@ void main() {
 
           await camera.startVideoRecording();
 
-          videoRecordingStoppedListener(Event('stop'));
+          final List<String> capturedEvents = <String>[];
+          mockMediaRecorder.removeEventListener = (
+            String type,
+            EventListener? callback, [
+            JSAny? options,
+          ]) {
+            capturedEvents.add(type);
+          };
+
+          videoRecordingStoppedListener.callAsFunction(null, Event('stop'));
 
           await Future<void>.microtask(() {});
 
-          verify(
-            () => mediaRecorder.removeEventListener('dataavailable', any()),
-          ).called(1);
+          expect(
+            capturedEvents.where((String e) => e == 'dataavailable').length,
+            1,
+          );
         });
 
         testWidgets('stops listening to the media recorder stop events',
@@ -1342,28 +1409,41 @@ void main() {
 
           await camera.startVideoRecording();
 
-          videoRecordingStoppedListener(Event('stop'));
+          final List<String> capturedEvents = <String>[];
+          mockMediaRecorder.removeEventListener = (
+            String type,
+            EventListener? callback, [
+            JSAny? options,
+          ]) {
+            capturedEvents.add(type);
+          };
+
+          videoRecordingStoppedListener.callAsFunction(null, Event('stop'));
 
           await Future<void>.microtask(() {});
 
-          verify(
-            () => mediaRecorder.removeEventListener('stop', any()),
-          ).called(1);
+          expect(
+            capturedEvents.where((String e) => e == 'stop').length,
+            1,
+          );
         });
 
         testWidgets('stops listening to the media recorder errors',
             (WidgetTester tester) async {
           final StreamController<ErrorEvent> onErrorStreamController =
               StreamController<ErrorEvent>();
+          final MockEventStreamProvider<Event> provider =
+              MockEventStreamProvider<Event>();
 
           final Camera camera = Camera(
             textureId: 1,
             cameraService: cameraService,
           )
             ..mediaRecorder = mediaRecorder
-            ..isVideoTypeSupported = isVideoTypeSupported;
+            ..isVideoTypeSupported = isVideoTypeSupported
+            ..mediaRecorderOnErrorProvider = provider;
 
-          when(() => mediaRecorder.onError)
+          when(() => provider.forTarget(mediaRecorder))
               .thenAnswer((_) => onErrorStreamController.stream);
 
           await camera.initialize();
@@ -1371,7 +1451,7 @@ void main() {
 
           await camera.startVideoRecording();
 
-          videoRecordingStoppedListener(Event('stop'));
+          videoRecordingStoppedListener.callAsFunction(null, Event('stop'));
 
           await Future<void>.microtask(() {});
 
@@ -1452,9 +1532,9 @@ void main() {
             'when a video recording is created', (WidgetTester tester) async {
           const String supportedVideoType = 'video/webm';
 
-          final MockMediaRecorder mediaRecorder = MockMediaRecorder();
-          when(() => mediaRecorder.onError)
-              .thenAnswer((_) => const Stream<Event>.empty());
+          final MockMediaRecorder mockMediaRecorder = MockMediaRecorder();
+          final MediaRecorder mediaRecorder =
+              createJSInteropWrapper(mockMediaRecorder) as MediaRecorder;
 
           final Camera camera = Camera(
             textureId: 1,
@@ -1466,22 +1546,20 @@ void main() {
           await camera.initialize();
           await camera.play();
 
-          late void Function(Event) videoDataAvailableListener;
-          late void Function(Event) videoRecordingStoppedListener;
+          late EventListener videoDataAvailableListener;
+          late EventListener videoRecordingStoppedListener;
 
-          when(
-            () => mediaRecorder.addEventListener('dataavailable', any()),
-          ).thenAnswer((Invocation invocation) {
-            videoDataAvailableListener =
-                invocation.positionalArguments[1] as void Function(Event);
-          });
-
-          when(
-            () => mediaRecorder.addEventListener('stop', any()),
-          ).thenAnswer((Invocation invocation) {
-            videoRecordingStoppedListener =
-                invocation.positionalArguments[1] as void Function(Event);
-          });
+          mockMediaRecorder.addEventListener = (
+            String type,
+            EventListener? callback, [
+            JSAny? options,
+          ]) {
+            if (type == 'dataavailable') {
+              videoDataAvailableListener = callback!;
+            } else if (type == 'stop') {
+              videoRecordingStoppedListener = callback!;
+            }
+          };
 
           final StreamQueue<VideoRecordedEvent> streamQueue =
               StreamQueue<VideoRecordedEvent>(camera.onVideoRecordedEvent);
@@ -1490,12 +1568,15 @@ void main() {
 
           Blob? finalVideo;
           camera.blobBuilder = (List<Blob> blobs, String videoType) {
-            finalVideo = Blob(blobs, videoType);
+            finalVideo = Blob(blobs.toJS, BlobPropertyBag(type: videoType));
             return finalVideo!;
           };
 
-          videoDataAvailableListener(FakeBlobEvent(Blob(<Object>[])));
-          videoRecordingStoppedListener(Event('stop'));
+          videoDataAvailableListener.callAsFunction(
+            null,
+            createJSInteropWrapper(FakeBlobEvent(Blob(<JSAny>[].toJS))),
+          );
+          videoRecordingStoppedListener.callAsFunction(null, Event('stop'));
 
           expect(
             await streamQueue.next,
@@ -1543,7 +1624,7 @@ void main() {
           await camera.initialize();
 
           final List<MediaStreamTrack> videoTracks =
-              camera.stream!.getVideoTracks();
+              camera.stream!.getVideoTracks().toDart;
           final MediaStreamTrack defaultVideoTrack = videoTracks.first;
 
           defaultVideoTrack.dispatchEvent(Event('ended'));
@@ -1570,7 +1651,7 @@ void main() {
           await camera.initialize();
 
           final List<MediaStreamTrack> videoTracks =
-              camera.stream!.getVideoTracks();
+              camera.stream!.getVideoTracks().toDart;
           final MediaStreamTrack defaultVideoTrack = videoTracks.first;
 
           camera.stop();
@@ -1589,16 +1670,22 @@ void main() {
             'emits an ErrorEvent '
             'when the media recorder fails '
             'when recording a video', (WidgetTester tester) async {
-          final MockMediaRecorder mediaRecorder = MockMediaRecorder();
+          final MockMediaRecorder mockMediaRecorder = MockMediaRecorder();
+          final MediaRecorder mediaRecorder =
+              createJSInteropWrapper(mockMediaRecorder) as MediaRecorder;
           final StreamController<ErrorEvent> errorController =
               StreamController<ErrorEvent>();
+          final MockEventStreamProvider<Event> provider =
+              MockEventStreamProvider<Event>();
 
           final Camera camera = Camera(
             textureId: textureId,
             cameraService: cameraService,
-          )..mediaRecorder = mediaRecorder;
+          )
+            ..mediaRecorder = mediaRecorder
+            ..mediaRecorderOnErrorProvider = provider;
 
-          when(() => mediaRecorder.onError)
+          when(() => provider.forTarget(mediaRecorder))
               .thenAnswer((_) => errorController.stream);
 
           final StreamQueue<ErrorEvent> streamQueue =
