@@ -3,14 +3,17 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:html' as html;
+import 'dart:js_interop';
 import 'dart:ui';
 import 'dart:ui_web' as ui_web;
 
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:flutter/foundation.dart';
+import 'package:web/web.dart' as web;
+import 'package:web/web.dart';
 
 import 'camera_service.dart';
+import 'pkg_web_tweaks.dart';
 import 'types/types.dart';
 
 String _getViewType(int cameraId) => 'plugins.flutter.io/camera_$cameraId';
@@ -48,10 +51,6 @@ class Camera {
     this.recorderOptions = const (audioBitrate: null, videoBitrate: null),
   }) : _cameraService = cameraService;
 
-  // A torch mode constraint name.
-  // See: https://w3c.github.io/mediacapture-image/#dom-mediatracksupportedconstraints-torch
-  static const String _torchModeKey = 'torch';
-
   /// The texture id used to register the camera view.
   final int textureId;
 
@@ -63,16 +62,16 @@ class Camera {
 
   /// The video element that displays the camera stream.
   /// Initialized in [initialize].
-  late final html.VideoElement videoElement;
+  late final web.HTMLVideoElement videoElement;
 
   /// The wrapping element for the [videoElement] to avoid overriding
   /// the custom styles applied in [_applyDefaultVideoStyles].
   /// Initialized in [initialize].
-  late final html.DivElement divElement;
+  late final web.HTMLDivElement divElement;
 
   /// The camera stream displayed in the [videoElement].
   /// Initialized in [initialize] and [play], reset in [stop].
-  html.MediaStream? stream;
+  web.MediaStream? stream;
 
   /// The stream of the camera video tracks that have ended playing.
   ///
@@ -82,14 +81,14 @@ class Camera {
   ///
   /// MediaStreamTrack.onended:
   /// https://developer.mozilla.org/en-US/docs/Web/API/MediaStreamTrack/onended
-  Stream<html.MediaStreamTrack> get onEnded => onEndedController.stream;
+  Stream<web.MediaStreamTrack> get onEnded => onEndedController.stream;
 
   /// The stream controller for the [onEnded] stream.
   @visibleForTesting
-  final StreamController<html.MediaStreamTrack> onEndedController =
-      StreamController<html.MediaStreamTrack>.broadcast();
+  final StreamController<web.MediaStreamTrack> onEndedController =
+      StreamController<web.MediaStreamTrack>.broadcast();
 
-  StreamSubscription<html.Event>? _onEndedSubscription;
+  StreamSubscription<web.Event>? _onEndedSubscription;
 
   /// The stream of the camera video recording errors.
   ///
@@ -98,15 +97,20 @@ class Camera {
   ///
   /// MediaRecorder.error:
   /// https://developer.mozilla.org/en-US/docs/Web/API/MediaRecorder/error_event
-  Stream<html.ErrorEvent> get onVideoRecordingError =>
+  Stream<web.ErrorEvent> get onVideoRecordingError =>
       videoRecordingErrorController.stream;
+
+  /// The stream provider for [MediaRecorder] error events.
+  @visibleForTesting
+  EventStreamProvider<Event> mediaRecorderOnErrorProvider =
+      EventStreamProviders.errorMediaRecorderEvent;
 
   /// The stream controller for the [onVideoRecordingError] stream.
   @visibleForTesting
-  final StreamController<html.ErrorEvent> videoRecordingErrorController =
-      StreamController<html.ErrorEvent>.broadcast();
+  final StreamController<web.ErrorEvent> videoRecordingErrorController =
+      StreamController<web.ErrorEvent>.broadcast();
 
-  StreamSubscription<html.Event>? _onVideoRecordingErrorSubscription;
+  StreamSubscription<web.Event>? _onVideoRecordingErrorSubscription;
 
   /// The camera flash mode.
   @visibleForTesting
@@ -117,36 +121,37 @@ class Camera {
 
   /// The current browser window used to access media devices.
   @visibleForTesting
-  html.Window? window = html.window;
+  web.Window window = web.window;
 
   /// The recorder used to record a video from the camera.
   @visibleForTesting
-  html.MediaRecorder? mediaRecorder;
+  web.MediaRecorder? mediaRecorder;
 
   /// Whether the video of the given type is supported.
   @visibleForTesting
   bool Function(String) isVideoTypeSupported =
-      html.MediaRecorder.isTypeSupported;
+      (String type) => web.MediaRecorder.isTypeSupported(type);
 
   /// The list of consecutive video data files recorded with [mediaRecorder].
-  final List<html.Blob> _videoData = <html.Blob>[];
+  final List<web.Blob> _videoData = <web.Blob>[];
 
   /// Completes when the video recording is stopped/finished.
   Completer<XFile>? _videoAvailableCompleter;
 
   /// A data listener fired when a new part of video data is available.
-  void Function(html.Event)? _videoDataAvailableListener;
+  void Function(web.BlobEvent)? _videoDataAvailableListener;
 
   /// A listener fired when a video recording is stopped.
-  void Function(html.Event)? _videoRecordingStoppedListener;
+  void Function(web.Event)? _videoRecordingStoppedListener;
 
   /// A builder to merge a list of blobs into a single blob.
   @visibleForTesting
   // TODO(stuartmorgan): Remove this 'ignore' once we don't analyze using 2.10
   // any more. It's a false positive that is fixed in later versions.
   // ignore: prefer_function_declarations_over_variables
-  html.Blob Function(List<html.Blob> blobs, String type) blobBuilder =
-      (List<html.Blob> blobs, String type) => html.Blob(blobs, type);
+  web.Blob Function(List<web.Blob> blobs, String type) blobBuilder =
+      (List<web.Blob> blobs, String type) =>
+          web.Blob(blobs.toJS, web.BlobPropertyBag(type: type));
 
   /// The stream that emits a [VideoRecordedEvent] when a video recording is created.
   Stream<VideoRecordedEvent> get onVideoRecordedEvent =>
@@ -166,9 +171,9 @@ class Camera {
       cameraId: textureId,
     );
 
-    videoElement = html.VideoElement();
+    videoElement = web.HTMLVideoElement();
 
-    divElement = html.DivElement()
+    divElement = web.HTMLDivElement()
       ..style.setProperty('object-fit', 'cover')
       ..append(videoElement);
 
@@ -185,12 +190,14 @@ class Camera {
 
     _applyDefaultVideoStyles(videoElement);
 
-    final List<html.MediaStreamTrack> videoTracks = stream!.getVideoTracks();
+    final List<web.MediaStreamTrack> videoTracks =
+        stream!.getVideoTracks().toDart;
 
     if (videoTracks.isNotEmpty) {
-      final html.MediaStreamTrack defaultVideoTrack = videoTracks.first;
-
-      _onEndedSubscription = defaultVideoTrack.onEnded.listen((html.Event _) {
+      final web.MediaStreamTrack defaultVideoTrack = videoTracks.first;
+      _onEndedSubscription = EventStreamProviders.endedEvent
+          .forTarget(defaultVideoTrack)
+          .listen((web.Event _) {
         onEndedController.add(defaultVideoTrack);
       });
     }
@@ -207,7 +214,7 @@ class Camera {
       );
       videoElement.srcObject = stream;
     }
-    await videoElement.play();
+    await videoElement.play().toDart;
   }
 
   /// Pauses the camera stream on the current frame.
@@ -217,14 +224,15 @@ class Camera {
 
   /// Stops the camera stream and resets the camera source.
   void stop() {
-    final List<html.MediaStreamTrack> videoTracks = stream!.getVideoTracks();
+    final List<web.MediaStreamTrack> videoTracks =
+        stream!.getVideoTracks().toDart;
     if (videoTracks.isNotEmpty) {
       onEndedController.add(videoTracks.first);
     }
 
-    final List<html.MediaStreamTrack>? tracks = stream?.getTracks();
+    final List<web.MediaStreamTrack>? tracks = stream?.getTracks().toDart;
     if (tracks != null) {
-      for (final html.MediaStreamTrack track in tracks) {
+      for (final web.MediaStreamTrack track in tracks) {
         track.stop();
       }
     }
@@ -246,8 +254,9 @@ class Camera {
 
     final int videoWidth = videoElement.videoWidth;
     final int videoHeight = videoElement.videoHeight;
-    final html.CanvasElement canvas =
-        html.CanvasElement(width: videoWidth, height: videoHeight);
+    final web.HTMLCanvasElement canvas = web.HTMLCanvasElement()
+      ..width = videoWidth
+      ..height = videoHeight;
     final bool isBackCamera = getLensDirection() == CameraLensDirection.back;
 
     // Flip the picture horizontally if it is not taken from a back camera.
@@ -257,16 +266,28 @@ class Camera {
         ..scale(-1, 1);
     }
 
-    canvas.context2D
-        .drawImageScaled(videoElement, 0, 0, videoWidth, videoHeight);
+    canvas.context2D.drawImageScaled(
+      videoElement,
+      0,
+      0,
+      videoWidth.toDouble(),
+      videoHeight.toDouble(),
+    );
 
-    final html.Blob blob = await canvas.toBlob('image/jpeg');
+    final Completer<web.Blob> blobCompleter = Completer<web.Blob>();
+    canvas.toBlob(
+      (web.Blob blob) {
+        blobCompleter.complete(blob);
+      }.toJS,
+      'image/jpeg',
+    );
+    final web.Blob blob = await blobCompleter.future;
 
     if (shouldEnableTorchMode) {
       _setTorchMode(enabled: false);
     }
 
-    return XFile(html.Url.createObjectUrl(blob));
+    return XFile(web.URL.createObjectURL(blob));
   }
 
   /// Returns a size of the camera video based on its first video track size.
@@ -274,25 +295,23 @@ class Camera {
   /// Returns [Size.zero] if the camera is missing a video track or
   /// the video track does not include the width or height setting.
   Size getVideoSize() {
-    final List<html.MediaStreamTrack> videoTracks =
-        videoElement.srcObject?.getVideoTracks() ?? <html.MediaStreamTrack>[];
+    final List<web.MediaStreamTrack> videoTracks =
+        (videoElement.srcObject as web.MediaStream?)?.getVideoTracks().toDart ??
+            <web.MediaStreamTrack>[];
 
     if (videoTracks.isEmpty) {
       return Size.zero;
     }
 
-    final html.MediaStreamTrack defaultVideoTrack = videoTracks.first;
-    final Map<dynamic, dynamic> defaultVideoTrackSettings =
+    final web.MediaStreamTrack defaultVideoTrack = videoTracks.first;
+
+    final web.MediaTrackSettings defaultVideoTrackSettings =
         defaultVideoTrack.getSettings();
 
-    final double? width = defaultVideoTrackSettings['width'] as double?;
-    final double? height = defaultVideoTrackSettings['height'] as double?;
+    final int width = defaultVideoTrackSettings.width;
+    final int height = defaultVideoTrackSettings.height;
 
-    if (width != null && height != null) {
-      return Size(width, height);
-    } else {
-      return Size.zero;
-    }
+    return Size(width.toDouble(), height.toDouble());
   }
 
   /// Sets the camera flash mode to [mode] by modifying the camera
@@ -307,13 +326,11 @@ class Camera {
   /// Throws a [CameraWebException] if the torch mode is not supported
   /// or the camera has not been initialized or started.
   void setFlashMode(FlashMode mode) {
-    final html.MediaDevices? mediaDevices = window?.navigator.mediaDevices;
-    final Map<dynamic, dynamic>? supportedConstraints =
-        mediaDevices?.getSupportedConstraints();
-    final bool torchModeSupported =
-        supportedConstraints?[_torchModeKey] as bool? ?? false;
+    final web.MediaDevices mediaDevices = window.navigator.mediaDevices;
+    final web.MediaTrackSupportedConstraints supportedConstraints =
+        mediaDevices.getSupportedConstraints();
 
-    if (!torchModeSupported) {
+    if (!supportedConstraints.torch) {
       throw CameraWebException(
         textureId,
         CameraErrorCode.torchModeNotSupported,
@@ -333,23 +350,20 @@ class Camera {
   /// Throws a [CameraWebException] if the torch mode is not supported
   /// or the camera has not been initialized or started.
   void _setTorchMode({required bool enabled}) {
-    final List<html.MediaStreamTrack> videoTracks =
-        stream?.getVideoTracks() ?? <html.MediaStreamTrack>[];
+    final List<web.MediaStreamTrack> videoTracks =
+        stream?.getVideoTracks().toDart ?? <web.MediaStreamTrack>[];
 
     if (videoTracks.isNotEmpty) {
-      final html.MediaStreamTrack defaultVideoTrack = videoTracks.first;
-
+      final web.MediaStreamTrack defaultVideoTrack = videoTracks.first;
       final bool canEnableTorchMode =
-          defaultVideoTrack.getCapabilities()[_torchModeKey] as bool? ?? false;
+          defaultVideoTrack.getCapabilities().torch.toDart.first.toDart;
 
       if (canEnableTorchMode) {
-        defaultVideoTrack.applyConstraints(<String, Object>{
-          'advanced': <Object>[
-            <String, Object>{
-              _torchModeKey: enabled,
-            }
-          ]
-        });
+        defaultVideoTrack.applyConstraints(
+          NonStandardFieldsOnMediaTrackConstraints.construct(
+            torch: enabled.toJS,
+          ),
+        );
       } else {
         throw CameraWebException(
           textureId,
@@ -397,13 +411,9 @@ class Camera {
       );
     }
 
-    zoomLevelCapability.videoTrack.applyConstraints(<String, Object>{
-      'advanced': <Object>[
-        <String, Object>{
-          ZoomLevelCapability.constraintName: zoom,
-        }
-      ]
-    });
+    zoomLevelCapability.videoTrack.applyConstraints(
+      NonStandardFieldsOnMediaTrackConstraints.construct(zoom: zoom.toJS),
+    );
   }
 
   /// Returns a lens direction of this camera.
@@ -411,19 +421,19 @@ class Camera {
   /// Returns null if the camera is missing a video track or
   /// the video track does not include the facing mode setting.
   CameraLensDirection? getLensDirection() {
-    final List<html.MediaStreamTrack> videoTracks =
-        videoElement.srcObject?.getVideoTracks() ?? <html.MediaStreamTrack>[];
+    final List<web.MediaStreamTrack> videoTracks =
+        (videoElement.srcObject as web.MediaStream?)?.getVideoTracks().toDart ??
+            <web.MediaStreamTrack>[];
 
     if (videoTracks.isEmpty) {
       return null;
     }
 
-    final html.MediaStreamTrack defaultVideoTrack = videoTracks.first;
-    final Map<dynamic, dynamic> defaultVideoTrackSettings =
+    final web.MediaStreamTrack defaultVideoTrack = videoTracks.first;
+    final web.MediaTrackSettings defaultVideoTrackSettings =
         defaultVideoTrack.getSettings();
 
-    final String? facingMode =
-        defaultVideoTrackSettings['facingMode'] as String?;
+    final String? facingMode = defaultVideoTrackSettings.facingModeNullable;
 
     if (facingMode != null) {
       return _cameraService.mapFacingModeToLensDirection(facingMode);
@@ -435,65 +445,65 @@ class Camera {
   /// Returns the registered view type of the camera.
   String getViewType() => _getViewType(textureId);
 
-  /// Starts a new video recording using [html.MediaRecorder].
+  /// Starts a new video recording using [web.MediaRecorder].
   ///
   /// Throws a [CameraWebException] if the browser does not support any of the
   /// available video mime types from [_videoMimeType].
   Future<void> startVideoRecording() async {
+    final web.MediaRecorderOptions options =
+        web.MediaRecorderOptions(mimeType: _videoMimeType);
+    if (recorderOptions.audioBitrate != null) {
+      options.audioBitsPerSecond = recorderOptions.audioBitrate!;
+    }
+    if (recorderOptions.videoBitrate != null) {
+      options.videoBitsPerSecond = recorderOptions.videoBitrate!;
+    }
+
     mediaRecorder ??=
-        html.MediaRecorder(videoElement.srcObject!, <String, Object>{
-      'mimeType': _videoMimeType,
-      if (recorderOptions.audioBitrate != null)
-        'audioBitsPerSecond': recorderOptions.audioBitrate!,
-      if (recorderOptions.videoBitrate != null)
-        'videoBitsPerSecond': recorderOptions.videoBitrate!,
-    });
+        web.MediaRecorder(videoElement.srcObject! as web.MediaStream, options);
 
     _videoAvailableCompleter = Completer<XFile>();
 
     _videoDataAvailableListener =
-        (html.Event event) => _onVideoDataAvailable(event);
+        (web.BlobEvent event) => _onVideoDataAvailable(event);
 
     _videoRecordingStoppedListener =
-        (html.Event event) => _onVideoRecordingStopped(event);
+        (web.Event event) => _onVideoRecordingStopped(event);
 
     mediaRecorder!.addEventListener(
       'dataavailable',
-      _videoDataAvailableListener,
+      _videoDataAvailableListener?.toJS,
     );
 
     mediaRecorder!.addEventListener(
       'stop',
-      _videoRecordingStoppedListener,
+      _videoRecordingStoppedListener?.toJS,
     );
 
-    _onVideoRecordingErrorSubscription =
-        mediaRecorder!.onError.listen((html.Event event) {
-      final html.ErrorEvent error = event as html.ErrorEvent;
+    _onVideoRecordingErrorSubscription = mediaRecorderOnErrorProvider
+        .forTarget(mediaRecorder)
+        .listen((web.Event event) {
+      final web.ErrorEvent error = event as web.ErrorEvent;
       videoRecordingErrorController.add(error);
     });
 
     mediaRecorder!.start();
   }
 
-  void _onVideoDataAvailable(html.Event event) {
-    final html.Blob? blob = (event as html.BlobEvent).data;
-
+  void _onVideoDataAvailable(web.BlobEvent event) {
     // Append the recorded part of the video to the list of all video data files.
-    if (blob != null) {
-      _videoData.add(blob);
-    }
+    _videoData.add(event.data);
   }
 
-  Future<void> _onVideoRecordingStopped(html.Event event) async {
+  Future<void> _onVideoRecordingStopped(web.Event event) async {
     if (_videoData.isNotEmpty) {
       // Concatenate all video data files into a single blob.
       final String videoType = _videoData.first.type;
-      final html.Blob videoBlob = blobBuilder(_videoData, videoType);
+      final web.Blob videoBlob = blobBuilder(_videoData, videoType);
 
       // Create a file containing the video blob.
       final XFile file = XFile(
-        html.Url.createObjectUrl(videoBlob),
+        web.URL.createObjectURL(videoBlob),
         mimeType: _videoMimeType,
         name: videoBlob.hashCode.toString(),
       );
@@ -509,12 +519,12 @@ class Camera {
     // Clean up the media recorder with its event listeners and video data.
     mediaRecorder!.removeEventListener(
       'dataavailable',
-      _videoDataAvailableListener,
+      _videoDataAvailableListener?.toJS,
     );
 
     mediaRecorder!.removeEventListener(
       'stop',
-      _videoDataAvailableListener,
+      _videoDataAvailableListener?.toJS,
     );
 
     await _onVideoRecordingErrorSubscription?.cancel();
@@ -612,7 +622,7 @@ class Camera {
       );
 
   /// Applies default styles to the video [element].
-  void _applyDefaultVideoStyles(html.VideoElement element) {
+  void _applyDefaultVideoStyles(web.HTMLVideoElement element) {
     final bool isBackCamera = getLensDirection() == CameraLensDirection.back;
 
     // Flip the video horizontally if it is not taken from a back camera.
