@@ -5,111 +5,158 @@
 // ignore_for_file: avoid_implementing_value_types
 
 import 'dart:async';
-import 'dart:html';
+import 'dart:js_interop';
 import 'dart:ui';
 
 // ignore_for_file: implementation_imports
 import 'package:camera_web/src/camera.dart';
 import 'package:camera_web/src/camera_service.dart';
+import 'package:camera_web/src/pkg_web_tweaks.dart';
 import 'package:camera_web/src/shims/dart_js_util.dart';
 import 'package:camera_web/src/types/types.dart';
 import 'package:cross_file/cross_file.dart';
 import 'package:mocktail/mocktail.dart';
+// TODO(srujzs): This is exported in `package:web` 0.6.0. Remove this when it is available.
+import 'package:web/src/helpers/events/streams.dart';
+import 'package:web/web.dart' as web;
 
-class MockWindow extends Mock implements Window {}
+@JSExport()
+class MockWindow {
+  late web.Navigator navigator;
+  late web.Screen screen;
+  late web.Document document;
+}
 
-class MockScreen extends Mock implements Screen {}
+@JSExport()
+class MockScreen {
+  late web.ScreenOrientation orientation;
+}
 
-class MockScreenOrientation extends Mock implements ScreenOrientation {}
+@JSExport()
+class MockScreenOrientation {
+  JSPromise<JSAny?> Function(web.OrientationLockType orientation) lock =
+      (web.OrientationLockType orientation) => Future<void>.value().toJS;
+  late void Function() unlock;
+  late web.OrientationType type;
+}
 
-class MockDocument extends Mock implements Document {}
+@JSExport()
+class MockDocument {
+  web.Element? documentElement;
+}
 
-class MockElement extends Mock implements Element {}
+@JSExport()
+class MockElement {
+  JSPromise<JSAny?> Function([FullscreenOptions options]) requestFullscreen =
+      ([FullscreenOptions? options]) => Future<void>.value().toJS;
+}
 
-class MockNavigator extends Mock implements Navigator {}
+@JSExport()
+class MockNavigator {
+  late web.MediaDevices mediaDevices;
+}
 
-class MockMediaDevices extends Mock implements MediaDevices {}
+@JSExport()
+class MockMediaDevices {
+  late JSPromise<web.MediaStream> Function([
+    web.MediaStreamConstraints? constraints,
+  ]) getUserMedia;
+
+  late web.MediaTrackSupportedConstraints Function() getSupportedConstraints;
+  late JSPromise<JSArray<web.MediaDeviceInfo>> Function() enumerateDevices;
+}
 
 class MockCameraService extends Mock implements CameraService {}
 
-class MockMediaStreamTrack extends Mock implements MediaStreamTrack {}
+@JSExport()
+class MockMediaStreamTrack {
+  late web.MediaTrackCapabilities Function() getCapabilities;
+  web.MediaTrackSettings Function() getSettings =
+      () => web.MediaTrackSettings();
+  late JSPromise<JSAny?> Function([web.MediaTrackConstraints? constraints])
+      applyConstraints;
+
+  void Function() stop = () {};
+}
 
 class MockCamera extends Mock implements Camera {}
 
 class MockCameraOptions extends Mock implements CameraOptions {}
 
-class MockVideoElement extends Mock implements VideoElement {}
+@JSExport()
+class MockVideoElement {
+  web.MediaProvider? srcObject;
+  web.MediaError? error;
+}
 
 class MockXFile extends Mock implements XFile {}
 
 class MockJsUtil extends Mock implements JsUtil {}
 
-class MockMediaRecorder extends Mock implements MediaRecorder {}
+@JSExport()
+class MockMediaRecorder {
+  void Function(
+    String type,
+    web.EventListener? callback, [
+    JSAny options,
+  ]) addEventListener = (
+    String type,
+    web.EventListener? callback, [
+    JSAny? options,
+  ]) {};
+
+  void Function(
+    String type,
+    web.EventListener? callback, [
+    JSAny options,
+  ]) removeEventListener = (
+    String type,
+    web.EventListener? callback, [
+    JSAny? options,
+  ]) {};
+
+  void Function([int timeslice]) start = ([int? timeslice]) {};
+  void Function() pause = () {};
+  void Function() resume = () {};
+  void Function() stop = () {};
+
+  web.RecordingState state = 'inactive';
+}
 
 /// A fake [MediaStream] that returns the provided [_videoTracks].
-class FakeMediaStream extends Fake implements MediaStream {
+@JSExport()
+class FakeMediaStream {
   FakeMediaStream(this._videoTracks);
 
-  final List<MediaStreamTrack> _videoTracks;
+  final List<web.MediaStreamTrack> _videoTracks;
 
-  @override
-  List<MediaStreamTrack> getVideoTracks() => _videoTracks;
+  List<web.MediaStreamTrack> getVideoTracks() => _videoTracks;
 }
 
 /// A fake [MediaDeviceInfo] that returns the provided [_deviceId], [_label] and [_kind].
-class FakeMediaDeviceInfo extends Fake implements MediaDeviceInfo {
-  FakeMediaDeviceInfo(this._deviceId, this._label, this._kind);
+@JSExport()
+class FakeMediaDeviceInfo {
+  FakeMediaDeviceInfo(this.deviceId, this.label, this.kind);
 
-  final String _deviceId;
-  final String _label;
-  final String _kind;
-
-  @override
-  String? get deviceId => _deviceId;
-
-  @override
-  String? get label => _label;
-
-  @override
-  String? get kind => _kind;
+  final String deviceId;
+  final String label;
+  final String kind;
 }
 
 /// A fake [MediaError] that returns the provided error [_code] and [_message].
-class FakeMediaError extends Fake implements MediaError {
+@JSExport()
+class FakeMediaError {
   FakeMediaError(
-    this._code, [
-    String message = '',
-  ]) : _message = message;
+    this.code, [
+    this.message = '',
+  ]);
 
-  final int _code;
-  final String _message;
-
-  @override
-  int get code => _code;
-
-  @override
-  String? get message => _message;
-}
-
-/// A fake [DomException] that returns the provided error [_name] and [_message].
-class FakeDomException extends Fake implements DomException {
-  FakeDomException(
-    this._name, [
-    String? message,
-  ]) : _message = message;
-
-  final String _name;
-  final String? _message;
-
-  @override
-  String get name => _name;
-
-  @override
-  String? get message => _message;
+  final int code;
+  final String message;
 }
 
 /// A fake [ElementStream] that listens to the provided [_stream] on [listen].
-class FakeElementStream<T extends Event> extends Fake
+class FakeElementStream<T extends web.Event> extends Fake
     implements ElementStream<T> {
   FakeElementStream(this._stream);
 
@@ -128,31 +175,23 @@ class FakeElementStream<T extends Event> extends Fake
 }
 
 /// A fake [BlobEvent] that returns the provided blob [data].
-class FakeBlobEvent extends Fake implements BlobEvent {
-  FakeBlobEvent(this._blob);
+@JSExport()
+class FakeBlobEvent {
+  FakeBlobEvent(this.data);
 
-  final Blob? _blob;
-
-  @override
-  Blob? get data => _blob;
+  final web.Blob? data;
 }
 
 /// A fake [DomException] that returns the provided error [_name] and [_message].
-class FakeErrorEvent extends Fake implements ErrorEvent {
+@JSExport()
+class FakeErrorEvent {
   FakeErrorEvent(
-    String type, [
-    String? message,
-  ])  : _type = type,
-        _message = message;
+    this.type, [
+    this.message = '',
+  ]);
 
-  final String _type;
-  final String? _message;
-
-  @override
-  String get type => _type;
-
-  @override
-  String? get message => _message;
+  final String type;
+  final String message;
 }
 
 /// Returns a video element with a blank stream of size [videoSize].
@@ -162,14 +201,17 @@ class FakeErrorEvent extends Fake implements ErrorEvent {
 /// final videoElement = getVideoElementWithBlankStream(Size(100, 100));
 /// final videoStream = videoElement.captureStream();
 /// ```
-VideoElement getVideoElementWithBlankStream(Size videoSize) {
-  final CanvasElement canvasElement = CanvasElement(
-    width: videoSize.width.toInt(),
-    height: videoSize.height.toInt(),
-  )..context2D.fillRect(0, 0, videoSize.width, videoSize.height);
+web.HTMLVideoElement getVideoElementWithBlankStream(Size videoSize) {
+  final web.HTMLCanvasElement canvasElement = web.HTMLCanvasElement()
+    ..width = videoSize.width.toInt()
+    ..height = videoSize.height.toInt()
+    ..context2D.fillRect(0, 0, videoSize.width, videoSize.height);
 
-  final VideoElement videoElement = VideoElement()
+  final web.HTMLVideoElement videoElement = web.HTMLVideoElement()
     ..srcObject = canvasElement.captureStream();
 
   return videoElement;
 }
+
+class MockEventStreamProvider<T extends web.Event> extends Mock
+    implements web.EventStreamProvider<T> {}
