@@ -2,9 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'dart:html' as html;
-
 import 'package:flutter/foundation.dart';
+import 'package:web/web.dart' as web;
 
 /// The possible range of values for the zoom level configurable
 /// on the camera video track.
@@ -30,7 +29,7 @@ class ZoomLevelCapability {
   final double maximum;
 
   /// The video track capable of configuring the zoom level.
-  final html.MediaStreamTrack videoTrack;
+  final web.MediaStreamTrack videoTrack;
 
   @override
   bool operator ==(Object other) {
